some fixes for SwiftUI example
