+++ Example/KVKCalendar/ViewController.swift
@@ -12,7 +12,11 @@ import EventKit
 
 final class ViewController: UIViewController, KVKCalendarSettings {
     
-    var events = [Event]()
+    var events = [Event]() {
+        didSet {
+            calendarView.reloadData()
+        }
+    }
     var selectDate = Date()
     var style: Style {
         createCalendarStyle()
@@ -73,7 +77,6 @@ final class ViewController: UIViewController, KVKCalendarSettings {
         loadEvents(dateFormat: style.timeSystem.format) { (events) in
             DispatchQueue.main.asyncAfter(deadline: .now() + 3) { [weak self] in
                 self?.events = events
-                self?.calendarView.reloadData()
             }
         }
     }
@@ -107,11 +110,10 @@ final class ViewController: UIViewController, KVKCalendarSettings {
         // to track changing windows and theme of device
         
         loadEvents(dateFormat: style.timeSystem.format) { [weak self] (events) in
-            self?.events = events
             if let style = self?.style {
                 self?.calendarView.updateStyle(style)
             }
-            self?.calendarView.reloadData()
+            self?.events = events
         }
     }
 }
@@ -120,21 +122,8 @@ final class ViewController: UIViewController, KVKCalendarSettings {
 
 extension ViewController: CalendarDelegate {
     func didChangeEvent(_ event: Event, start: Date?, end: Date?) {
-        var eventTemp = event
-        guard let startTemp = start, let endTemp = end else { return }
-        
-        let startTime = timeFormatter(date: startTemp, format: style.timeSystem.format)
-        let endTime = timeFormatter(date: endTemp, format: style.timeSystem.format)
-        eventTemp.start = startTemp
-        eventTemp.end = endTemp
-        eventTemp.title = TextEvent(timeline: "\(startTime) - \(endTime)\n new time",
-                                    month: "\(startTime) - \(endTime)\n new time",
-                                    list: "\(startTime) - \(endTime)\n new time")
-        
-        if let idx = events.firstIndex(where: { $0.compare(eventTemp) }) {
-            events.remove(at: idx)
-            events.append(eventTemp)
-            calendarView.reloadData()
+        if let result = handleChangingEvent(event, start: start, end: end) {
+            events.replaceSubrange(result.range, with: result.events)
         }
     }
     
@@ -166,20 +155,9 @@ extension ViewController: CalendarDelegate {
     }
     
     func didAddNewEvent(_ event: Event, _ date: Date?) {
-        var newEvent = event
-        
-        guard let start = date, let end = Calendar.current.date(byAdding: .minute, value: 30, to: start) else { return }
-        
-        let startTime = timeFormatter(date: start, format: style.timeSystem.format)
-        let endTime = timeFormatter(date: end, format: style.timeSystem.format)
-        newEvent.start = start
-        newEvent.end = end
-        newEvent.ID = "\(events.count + 1)"
-        newEvent.title = TextEvent(timeline: "\(startTime) - \(endTime)\n new time",
-                                   month: "\(startTime) - \(endTime)\n new time",
-                                   list: "\(startTime) - \(endTime)\n new time")
-        events.append(newEvent)
-        calendarView.reloadData()
+        if let newEvent = handleNewEvent(event, date: date) {
+            events.append(newEvent)
+        }
     }
 }
 
@@ -198,58 +176,20 @@ extension ViewController: CalendarDataSource {
     
     @available(iOS 14.0, *)
     func willDisplayEventOptionMenu(_ event: Event, type: CalendarType) -> (menu: UIMenu, customButton: UIButton?)? {
-        guard type == .day else { return nil }
-        
-        let action = UIAction(title: "Test", attributes: .destructive) { _ in
-            print("test tap")
-        }
-        
-        return (UIMenu(title: "Test menu", children: [action]), nil)
+        handleOptionMenu(type: type)
     }
     
     func eventsForCalendar(systemEvents: [EKEvent]) -> [Event] {
         // if you want to get a system events, you need to set style.systemCalendars = ["test"]
-        let mappedEvents = systemEvents.compactMap { (event) -> Event in
-            let startTime = timeFormatter(date: event.startDate, format: style.timeSystem.format)
-            let endTime = timeFormatter(date: event.endDate, format: style.timeSystem.format)
-            event.title = "\(startTime) - \(endTime)\n\(event.title ?? "")"
-            
-            return Event(event: event)
-        }
-        
-        return events + mappedEvents
+        handleEvents(systemEvents: systemEvents)
     }
     
     func willDisplayEventView(_ event: Event, frame: CGRect, date: Date?) -> EventViewGeneral? {
-        guard event.ID == "2" else { return nil }
-        
-        return CustomViewEvent(style: calendarView.style, event: event, frame: frame)
+        handleCustomEventView(event: event, style: style, frame: frame)
     }
     
     func dequeueCell<T>(dateParameter: DateParameter, type: CalendarType, view: T, indexPath: IndexPath) -> KVKCalendarCellProtocol? where T: UIScrollView {
-        switch type {
-        case .year where dateParameter.date?.month == Date().month:
-            let cell = (view as? UICollectionView)?.kvkDequeueCell(indexPath: indexPath) { (cell: CustomDayCell) in
-                cell.imageView.image = UIImage(named: "ic_stub")
-            }
-            return cell
-        case .day, .week, .month:
-            guard dateParameter.date?.day == Date().day && dateParameter.type != .empty else { return nil }
-            
-            let cell = (view as? UICollectionView)?.kvkDequeueCell(indexPath: indexPath) { (cell: CustomDayCell) in
-                cell.imageView.image = UIImage(named: "ic_stub")
-            }
-            return cell
-        case .list:
-            guard dateParameter.date?.day == 14 else { return nil }
-            
-            let cell = (view as? UITableView)?.kvkDequeueCell { (cell) in
-                cell.backgroundColor = .systemRed
-            }
-            return cell
-        default:
-            return nil
-        }
+        handleCell(dateParameter: dateParameter, type: type, view: view, indexPath: indexPath)
     }
     
     func willDisplayEventViewer(date: Date, frame: CGRect) -> UIView? {
@@ -258,16 +198,7 @@ extension ViewController: CalendarDataSource {
     }
     
     func sizeForCell(_ date: Date?, type: CalendarType) -> CGSize? {
-        guard type == .month && UIDevice.current.userInterfaceIdiom == .phone else { return nil }
-        
-        switch calendarView.style.month.scrollDirection {
-        case .vertical:
-            return CGSize(width: view.bounds.width / 7, height: 70)
-        case .horizontal:
-            return nil
-        @unknown default:
-            return nil
-        }
+        handleSizeCell(type: type, stye: calendarView.style, view: view)
     }
 }
